+++ src/sys/class/alo/exception/abstractexception.php
@@ -2,8 +2,6 @@
 
    namespace Alo\Exception;
 
-   use Exception;
-
    if (!defined('GEN_START')) {
       http_response_code(404);
       die();
@@ -14,15 +12,15 @@
     *
     * @author Art <a.molcanovas@gmail.com>
     */
-   abstract class AbstractException extends Exception {
+   abstract class AbstractException extends \Exception {
 
       /**
        * Creates the exception
        *
        * @author Art <a.molcanovas@gmail.com>
-       * @param string    $message  Exception message
-       * @param int       $code     Exception code
-       * @param Exception $previous Previous exception, if chaining
+       * @param string     $message  Exception message
+       * @param int        $code     Exception code
+       * @param \Exception $previous Previous exception, if chaining
        */
       function __construct($message = '', $code = 0, $previous = null) {
          parent::__construct($message, $code, $previous);