+++ changelog.md
@@ -2,4 +2,4 @@
 • Trait support added - see app/traits
 • Kint external library updated
 • Slashes in paths replaced with DIRECTORY_SEPARATOR
-• Again
+• Profiling class added