+++ changelog.md
@@ -4,3 +4,4 @@
 • Slashes in paths replaced with DIRECTORY_SEPARATOR
 • Profiling class added
 • The Router now has a lot more getters
+• Error, autoloading and exception handlers moved to \Alo\Handler