+++ changelog.md
@@ -7,3 +7,4 @@
 * The Router now has a lot more getters
 * Error, autoloading and exception handlers moved to \Alo\Handler
 * Committed function tester (forgot about it earlier)
+* Fixed an error where any uncaught exception would force the 404 error page.