+++ changelog.md
@@ -9,3 +9,4 @@
 * Committed function tester (forgot about it earlier)
 * Fixed an error where any uncaught exception would force the 404 error page.
 * Many more statics added to \Alo
+* Error divs prettified